Change node structure (see description)
DESCRIPTION

Additions:
    1) imgproc.cpp is a new (separated) node, which subscribes to /in/image
and publishes data in /out/image. Future C++ code for OpenCV-based algorithms
should be in this file.
    2) webcam.py is a new node, which connects to Android webcam server
and publishes image stream to /in/image. It can be used as a source of
images (instead of drone camera) for easier debugging of image
processing algorithms.
    3) new .launch file was added

Changes:
    1) controller.py does not process images anymore. It will be mostly
rewritten in the following commits